src/backend/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
Django==4.2
django-ninja==1.3.0
djangorestframework==3.15.2
psycopg2==2.9.9
pydantic==2.9.2
pydantic_core==2.23.4
sqlparse==0.5.1
typing_extensions==4.12.2
